requirements.txt
discord.py==1.5.1
dnspython==2.0.0
pymongo==3.11.1
pillow==8.0.1
deeppyer==1.0.1
async-timeout==3.0.1
requests==2.25.0
praw==7.1.0
hentai==3.1.3
asyncio==3.4.3
giphy-client==1.0.0
tweepy==3.9.0
mal-api==0.3.1
malclient==0.1.5
COVID19Py==0.3.0
pdf2image==1.14.0
imgurpython==1.1.7
numpy==1.19.4
opencv-python-headless==4.4.0.46
randfacts==0.2.4
python-dotenv==0.15.0
rule34==1.7.4
aiohttp
ratelimiter